+++ EssentialFeed/EssentialFeediOS/Feed UI/Composers/FeedViewAdapter.swift
@@ -1,0 +1,32 @@
+//
+//  FeedViewAdapter.swift
+//  EssentialFeediOS
+//
+//  Created by Stoyan Kostov on 28.06.23.
+//
+
+import UIKit
+
+final class FeedViewAdapter: FeedView {
+    private weak var controller: FeedViewController?
+    private let imageLoader: FeedImageDataLoader
+
+    init(controller: FeedViewController, imageLoader: FeedImageDataLoader) {
+        self.controller = controller
+        self.imageLoader = imageLoader
+    }
+
+    func display(_ viewModel: FeedViewModel) {
+        controller?.tableModel = viewModel.feed.map { model in
+            let adapter = FeedImageDataLoaderPresentationAdapter<WeakRefVirtualProxy<FeedImageCellController>, UIImage>(model: model, imageLoader: imageLoader)
+            let view = FeedImageCellController(delegate: adapter)
+
+            adapter.presenter = FeedImagePresenter(
+                view: WeakRefVirtualProxy(view),
+                imageTransformer: UIImage.init
+            )
+
+            return view
+        }
+    }
+}
